move loading of the recipes to the context
