+++ src/context/RecipeContext.tsx
@@ -1,6 +1,8 @@
-import React, { createContext, useReducer } from 'react';
+import React, { createContext, useReducer, useEffect } from 'react';
 import recipeReducer from './recipeReducer.tsx';
 import { Recipe, RecipeAction } from './types';
+import { RecipeActionTypes } from "../context/types.tsx";
+import { getRecipes } from "../context/RecipesService.tsx";
 
 export interface RecipeContextInterface {
     recipes: Recipe[];
@@ -17,6 +19,20 @@ export const RecipeContext = createContext<RecipeContextInterface>({
 
 export const RecipeProvider = ({ children }) => {
     const [recipes, dispatch] = useReducer(recipeReducer, []);
+
+    useEffect(() => {
+        const fetchRecipes = async () => {
+            try {
+                let payload = await getRecipes()
+                console.log("(on context) payload", payload);
+                dispatch({ type: RecipeActionTypes.FETCH_RECIPES, payload: payload });
+            } catch (error) {
+                dispatch({ type: RecipeActionTypes.ERROR, payload: "Something went wrong" });
+            }
+        };
+        console.log("(on cotext) fetchRecipes");
+        fetchRecipes();
+    }, [dispatch]);
 
     return (
         <RecipeContext.Provider value={{ recipes, dispatch }}>
